refactor(docs-infra): Move router-specific app configuration to separate file
This moves the router configuration for the app to a separate file since
there is a lot of it.
