+++ adev/src/app/app.config.ts
@@ -26,56 +26,21 @@ import {
   windowProvider,
 } from '@angular/docs';
 import {provideClientHydration} from '@angular/platform-browser';
-import {
-  RouteReuseStrategy,
-  Router,
-  TitleStrategy,
-  createUrlTreeFromSnapshot,
-  provideRouter,
-  withComponentInputBinding,
-  withInMemoryScrolling,
-  withViewTransitions,
-} from '@angular/router';
 import environment from './environment';
 import {PREVIEWS_COMPONENTS_MAP} from './../assets/previews/previews';
-import {ADevTitleStrategy} from './core/services/a-dev-title-strategy';
 import {AnalyticsService} from './core/services/analytics/analytics.service';
 import {ContentLoader} from './core/services/content-loader.service';
 import {CustomErrorHandler} from './core/services/errors-handling/error-handler';
 import {ExampleContentLoader} from './core/services/example-content-loader.service';
-import {ReuseTutorialsRouteStrategy} from './features/tutorial/tutorials-route-reuse-strategy';
-import {routes} from './routes';
 import {CURRENT_MAJOR_VERSION} from './core/providers/current-version';
-import {AppScroller} from './app-scroller';
+import {routerProviders} from './router_providers';
 
 export const appConfig: ApplicationConfig = {
   providers: [
-    provideRouter(
-      routes,
-      withInMemoryScrolling(),
-      withViewTransitions({
-        onViewTransitionCreated: ({transition, to}) => {
-          const router = inject(Router);
-          const toTree = createUrlTreeFromSnapshot(to, []);
-          // Skip the transition if the only thing changing is the fragment and queryParams
-          if (
-            router.isActive(toTree, {
-              paths: 'exact',
-              matrixParams: 'exact',
-              fragment: 'ignored',
-              queryParams: 'ignored',
-            })
-          ) {
-            transition.skipTransition();
-          }
-        },
-      }),
-      withComponentInputBinding(),
-    ),
+    routerProviders,
     provideExperimentalZonelessChangeDetection(),
     provideClientHydration(),
     provideHttpClient(withFetch()),
-    provideEnvironmentInitializer(() => inject(AppScroller)),
     provideEnvironmentInitializer(() => inject(AnalyticsService)),
     provideAlgoliaSearchClient(environment),
     {
@@ -88,14 +53,9 @@ export const appConfig: ApplicationConfig = {
     {provide: DOCS_CONTENT_LOADER, useClass: ContentLoader},
     {provide: EXAMPLE_VIEWER_CONTENT_LOADER, useClass: ExampleContentLoader},
     {
-      provide: RouteReuseStrategy,
-      useClass: ReuseTutorialsRouteStrategy,
-    },
-    {
       provide: WINDOW,
       useFactory: (document: Document) => windowProvider(document),
       deps: [DOCUMENT],
     },
-    {provide: TitleStrategy, useClass: ADevTitleStrategy},
   ],
 };
